feature: Add featured sections into their own component
