+++ src/pages/index.tsx
@@ -1,53 +1,9 @@
-import {
-  ArrowPathIcon,
-  CloudArrowUpIcon,
-  CogIcon,
-  LockClosedIcon,
-  ServerIcon,
-  ShieldCheckIcon,
-} from '@heroicons/react/24/outline';
 import { ArrowTopRightOnSquareIcon } from '@heroicons/react/20/solid';
 import LandingPageLayout from '../layouts/LandingPageLayout';
 import Hero from '@/ui/Hero';
+import FeaturedScreenshot from '@/ui/FeaturedScreenshot';
+import FeaturedGrid from '@/ui/FeaturedGrid';
 
-const features = [
-  {
-    name: 'Push to Deploy',
-    description:
-      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi vitae lobortis.',
-    icon: CloudArrowUpIcon,
-  },
-  {
-    name: 'SSL Certificates',
-    description:
-      'Qui aut temporibus nesciunt vitae dicta repellat sit dolores pariatur. Temporibus qui illum aut.',
-    icon: LockClosedIcon,
-  },
-  {
-    name: 'Simple Queues',
-    description:
-      'Rerum quas incidunt deleniti quaerat suscipit mollitia. Amet repellendus ut odit dolores qui.',
-    icon: ArrowPathIcon,
-  },
-  {
-    name: 'Advanced Security',
-    description:
-      'Ullam laboriosam est voluptatem maxime ut mollitia commodi. Et dignissimos suscipit perspiciatis.',
-    icon: ShieldCheckIcon,
-  },
-  {
-    name: 'Powerful API',
-    description:
-      'Ab a facere voluptatem in quia corrupti veritatis aliquam. Veritatis labore quaerat ipsum quaerat id.',
-    icon: CogIcon,
-  },
-  {
-    name: 'Database Backups',
-    description:
-      'Quia qui et est officia cupiditate qui consectetur. Ratione similique et impedit ea ipsum et.',
-    icon: ServerIcon,
-  },
-];
 const blogPosts = [
   {
     id: 1,
@@ -115,74 +71,8 @@ export default function Home() {
         <LandingPageLayout>
           <main>
             <Hero />
-            {/* Feature section with screenshot */}
-            <div className='relative pt-16 bg-gray-50 sm:pt-24 lg:pt-32'>
-              <div className='max-w-md px-6 mx-auto text-center sm:max-w-3xl lg:max-w-7xl lg:px-8'>
-                <div>
-                  <h2 className='text-lg font-semibold text-cyan-600'>
-                    Serverless
-                  </h2>
-                  <p className='mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl'>
-                    No server? No problem.
-                  </p>
-                  <p className='mx-auto mt-5 text-xl text-gray-500 max-w-prose'>
-                    Phasellus lorem quam molestie id quisque diam aenean nulla
-                    in. Accumsan in quis quis nunc, ullamcorper malesuada.
-                    Eleifend condimentum id viverra nulla.
-                  </p>
-                </div>
-                <div className='mt-12 -mb-10 sm:-mb-24 lg:-mb-80'>
-                  <img
-                    className='rounded-lg shadow-xl ring-1 ring-black ring-opacity-5'
-                    src='https://tailwindui.com/img/component-images/green-project-app-screenshot.jpg'
-                    alt=''
-                  />
-                </div>
-              </div>
-            </div>
-
-            {/* Feature section with grid */}
-            <div className='relative py-16 bg-white sm:py-24 lg:py-32'>
-              <div className='max-w-md px-6 mx-auto text-center sm:max-w-3xl lg:max-w-7xl lg:px-8'>
-                <h2 className='text-lg font-semibold text-cyan-600'>
-                  Deploy faster
-                </h2>
-                <p className='mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl'>
-                  Everything you need to deploy your app
-                </p>
-                <p className='mx-auto mt-5 text-xl text-gray-500 max-w-prose'>
-                  Phasellus lorem quam molestie id quisque diam aenean nulla in.
-                  Accumsan in quis quis nunc, ullamcorper malesuada. Eleifend
-                  condimentum id viverra nulla.
-                </p>
-                <div className='mt-12'>
-                  <div className='grid grid-cols-1 gap-8 sm:grid-cols-2 lg:grid-cols-3'>
-                    {features.map((feature) => (
-                      <div key={feature.name} className='pt-6'>
-                        <div className='flow-root px-6 pb-8 rounded-lg bg-gray-50'>
-                          <div className='-mt-6'>
-                            <div>
-                              <span className='inline-flex items-center justify-center p-3 rounded-md shadow-lg bg-gradient-to-r from-teal-500 to-cyan-600'>
-                                <feature.icon
-                                  className='w-6 h-6 text-white'
-                                  aria-hidden='true'
-                                />
-                              </span>
-                            </div>
-                            <h3 className='mt-8 text-lg font-medium tracking-tight text-gray-900'>
-                              {feature.name}
-                            </h3>
-                            <p className='mt-5 text-base text-gray-500'>
-                              {feature.description}
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            </div>
+            <FeaturedScreenshot />
+            <FeaturedGrid />
 
             {/* Testimonial section */}
             <div className='pb-16 bg-gradient-to-r from-teal-500 to-cyan-600 lg:relative lg:z-10 lg:pb-0'>
